Extract non-owning proxy + shapes projection out of useLayerData
Intermediate decomposition pass, no behaviour change. Moves two cohesive,
kind-local chunks out of the 1.6kloc hook:

- resolvers/nonOwningResolver.ts — `createNonOwningResolver` + its lifecycle
  rationale (the store-ownership exception for points).
- shapesProjection.ts — the shapes feature-state / fill-colour projection helpers
  and their cache-entry types (`ShapePrebuiltEntry`, `ShapeFillColorEntry`,
  `getStableShapeFeatureStateRuntime`, signature/serialise helpers). This is the
  `project()` half of ADR 0004 §4; a vis-local waypoint before Step 3 relocates it
  into @spatialdata/layers.

useLayerData.ts drops from 1623 to 1446 lines and imports both. The shapes read
path now lives in a small dedicated module, so Track B / Step 3 touch it rather than
the hook. Behaviour-preserving: control-char signature separators kept byte-identical;
full suite (550 tests) green, vis typecheck + build clean.
